refactor: modulize dev tools
