+++ packages/integrations/vue-devtools/functions/setupDevtools.ts
@@ -1,13 +1,14 @@
-import type { CustomInspectorNode } from '@vue/devtools-api'
 import { setupDevtoolsPlugin } from '@vue/devtools-api'
-import type { Exposition, ExpositionValues } from '@exposition/core'
+import type { Exposition } from '@exposition/core'
 
-import { expositionLabel, id, inspectorId, stateType, timelineId } from '../utils/config'
-import type { DevtoolsContext } from '../@types/api'
+import { expositionLabel, id, stateType } from '../utils/config'
+import type { DevtoolsContext, OnUpdateHandler } from '../@types/api'
 import { createSettingsViews } from '../views/createSettingsViews'
-import { actionLog, log } from '../utils/logs'
+import { createTimelineViews } from '../views/createTimelineViews'
+import { createInspectorViews } from '../views/createInspectorViews'
 import { defineDevtoolsSettings } from './defineDevtoolsSettings'
 import { defineExpositionState } from './defineExpositionState'
+import { updateState } from './updateState'
 
 /**
  * View debugger by running:
@@ -15,7 +16,7 @@ import { defineExpositionState } from './defineExpositionState'
  * _Do not forget to set the logging setting to `verbose` in your browser_
  */
 
-export function setupDevtools<T extends Exposition<any>>(app: any, options: { exposition: T; onUpdate: (exposition: ExpositionValues<T>, enabled: boolean) => void }) {
+export function setupDevtools<T extends Exposition<any>>(app: any, options: { exposition: T; onUpdate: OnUpdateHandler<T> }) {
   const settings = defineDevtoolsSettings()
   const state = defineExpositionState(options.exposition)
 
@@ -31,187 +32,17 @@ export function setupDevtools<T extends Exposition<any>>(app: any, options: { ex
       api,
       settings,
       state,
+      onUpdate: options.onUpdate,
     }
 
     function main(): void {
-      createSettingsViews(context, updateState)
+      createSettingsViews(context)
+      createTimelineViews(context)
+      createInspectorViews(context)
+
       state.loadFromStore()
-      updateState()
+      updateState(context)
     }
-
-    function updateState(): void {
-      log('Updating values to: %s', JSON.stringify(state.getValues()))
-
-      options.onUpdate(state.getValues(), settings.isEnabled('active'))
-
-      api.sendInspectorState(inspectorId)
-      api.sendInspectorTree(inspectorId)
-
-      api.addTimelineEvent({
-        layerId: timelineId,
-        event: {
-          title: 'updateExposition',
-          time: api.now(),
-          data: state.getValues(),
-        },
-      })
-    }
-
-    function createUpdateStateHandler(beforeUpdateHandler: Function = () => undefined): () => void {
-      return function () {
-        beforeUpdateHandler()
-        updateState()
-      }
-    }
-
-    api.addTimelineLayer({
-      id: timelineId,
-      label: expositionLabel,
-      color: 0x6004DB,
-    })
-
-    api.addInspector({
-      id: inspectorId,
-      label: 'Exposition',
-      icon: 'auto_stories',
-      actions: [
-        {
-          icon: 'restore',
-          tooltip: 'Reset the exposition to its initial state',
-          action: createUpdateStateHandler(() => {
-            actionLog('clicked restore action')
-            state.reset()
-          }),
-        },
-      ],
-    })
-
-    api.on.getInspectorTree((payload) => {
-      if (payload.inspectorId === inspectorId) {
-        const scenarioElements = Object.keys(state.value).reduce((accumulator, key) => {
-          const { id, value, initialValue } = state.value[key]
-
-          const isInitialValue = value === initialValue
-
-          return [
-            ...accumulator,
-            {
-              id,
-              label: id,
-              tags: [
-                isInitialValue
-                  ? {
-                      label: 'default',
-                      textColor: 0xFFFFFF,
-                      backgroundColor: 0x000000,
-                    }
-                  : { label: 'modified', textColor: 0xFFC495, backgroundColor: 0xAF4C05 },
-              ],
-            },
-          ]
-        }, [] as CustomInspectorNode[])
-
-        payload.rootNodes = [
-          {
-            id: 'settings',
-            label: '⚙ Settings',
-            tags: [
-              {
-                label: 'awesome',
-                textColor: 0xFFFFFF,
-                backgroundColor: 0x000000,
-              },
-            ],
-          },
-          {
-            id: 'scenarios',
-            label: '📕 Scenarios',
-            children: scenarioElements,
-          },
-        ]
-      }
-    })
-
-    api.on.getInspectorState((payload) => {
-      if (payload.inspectorId === inspectorId) {
-        if (payload.nodeId === 'scenarios') {
-          payload.state = {
-            // @ts-expect-error - figure out why TypeScript is angry
-            scenarios: Object.keys(state.getValues()).reduce((accumulator, key) => {
-              return [
-                ...accumulator,
-                {
-                  key,
-                  value: state.value[key].value,
-                },
-              ]
-            }, []),
-          }
-        }
-
-        if (Object.keys(state.value).includes(payload.nodeId)) {
-          const scenario = state.value[payload.nodeId]
-
-          payload.state = {
-            state: [
-              {
-                key: 'scenario',
-                value: scenario.id,
-              },
-              {
-                key: 'initialValue',
-                value: scenario.initialValue,
-              },
-              {
-                key: 'value',
-                value: {
-                  _custom: {
-                    type: null,
-                    value: scenario.value,
-                    actions: [{
-                      icon: 'restore',
-                      tooltip: 'Reset the value of the scenario',
-                      action: createUpdateStateHandler(() => {
-                        actionLog('restore scenario %s settings', scenario.id)
-                        // @ts-expect-error - Allow dynamic state definition in this case
-                        state.update({
-                          [scenario.id]: scenario.initialValue,
-                        })
-                      }),
-                    }],
-                  },
-                },
-                editable: false,
-              },
-            ],
-            // @ts-expect-error - figure out why TypeScript is angry
-            options: scenario.options.map((option, index) => {
-              return {
-                key: index,
-                value: {
-                  _custom: {
-                    type: null,
-                    value: option,
-                    actions: [{
-                      icon: 'check',
-                      tooltip: `Set "${option}" as the new value`,
-                      action: createUpdateStateHandler(() => {
-                        actionLog('set new value "%s" for scenario "%s"', option, scenario.id)
-                        // @ts-expect-error - Allow dynamic state definition in this case
-                        state.update({
-                          [scenario.id]: option,
-                        })
-                      }),
-                    }],
-                  },
-                },
-                editable: false,
-              }
-            }),
-          }
-        }
-      }
-    })
 
     main()
   })
